requirements.txt
casttube==0.2.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
funcy==1.15
idna==2.10
ifaddr==0.1.7
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
memoization==0.3.1
protobuf==3.14.0
PyChromecast==7.5.1
python-magic==0.4.18
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
zeroconf==0.28.6
